+++ src/geefetch/data/get.py
@@ -11,6 +11,7 @@ from rasterio.crs import CRS
 from retry import retry
 from rich.progress import Progress
 
+from ..cli.omegaconfig import SpeckleFilterConfig, TerrainNormalizationConfig
 from ..utils.enums import CompositeMethod, DType, Format, P2Orbit, ResamplingMethod, S1Orbit
 from ..utils.progress import default_bar
 from ..utils.rasterio import create_vrt
@@ -619,6 +620,8 @@ def download_s1(
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
+    speckle_filter_config: SpeckleFilterConfig | None = None,
+    terrain_normalization_config: TerrainNormalizationConfig | None = None,
     orbit: S1Orbit = S1Orbit.ASCENDING,
     resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
     tile_range: tuple[float, float] | None = None,
@@ -656,6 +659,10 @@ def download_s1(
         The data type of the downloaded images. Defaults to DType.Float32.
     filter_polygon : shapely.Geometry | None
         More fine-grained AOI than `bbox`. Defaults to None.
+    speckle_filter_config : SpeckleFilterConfig | None
+        speckle_filtering configurations
+    terrain_normalization_config: TerrainNormalizationConfig | None
+        terrain_normalization configurations
     orbit : S1Orbit
         The orbit used to filter Sentinel-1 images. Defaults to S1Orbit.ASCENDING.
     resampling : ResamplingMethod
@@ -703,6 +710,8 @@ def download_s1(
             "selected_bands": selected_bands,
             "resampling": resampling,
             "resolution": resolution,
+            "speckle_filter_config": speckle_filter_config,
+            "terrain_normalization_config": terrain_normalization_config,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
         tile_range=tile_range,