+++ src/geefetch/data/satellites/s1.py
@@ -1,13 +1,16 @@
 import logging
+from dataclasses import astuple
 from typing import Any
 
-import ee
 from ee.filter import Filter
 from ee.image import Image
 from ee.imagecollection import ImageCollection
+from gee_s1_processing.border_noise_correction import f_mask_edges
+from gee_s1_processing.wrapper import speckle_filter_wrapper, terrain_normalization_wrapper
 from geobbox import GeoBoundingBox
 from shapely import Polygon
 
+from ...cli.omegaconfig import SpeckleFilterConfig, TerrainNormalizationConfig
 from ...utils.enums import CompositeMethod, DType, ResamplingMethod, S1Orbit
 from ...utils.rasterio import WGS84
 from ..downloadables import DownloadableGeedimImage, DownloadableGeedimImageCollection
@@ -22,6 +25,8 @@ __all__ = ["S1"]
 class S1(SatelliteABC):
     _bands = ["HH", "HV", "VV", "VH", "angle"]
     _default_selected_bands = ["VV", "VH"]
+    speckle_filter_config: SpeckleFilterConfig | None
+    terrain_normalization_config: TerrainNormalizationConfig | None
 
     @property
     def bands(self) -> list[str]:
@@ -42,6 +47,12 @@ class S1(SatelliteABC):
     @property
     def resolution(self):
         return 10
+
+    @property
+    def is_preprocessed(self):
+        return (self.speckle_filter_config is not None) or (
+            self.terrain_normalization_config is not None
+        )
 
     def get_col(
         self,
@@ -95,7 +106,7 @@ class S1(SatelliteABC):
                 if band != "angle"
             ]
         )
-        col = ImageCollection("COPERNICUS/S1_GRD")
+        col = ImageCollection("COPERNICUS/S1_GRD_FLOAT")
         if start_date is not None and end_date is not None:
             col = col.filterDate(start_date, end_date)
         col = col.filterBounds(bounds).filter(band_filter).filter(Filter.eq("instrumentMode", "IW"))
@@ -107,7 +118,11 @@ class S1(SatelliteABC):
                 pass
             case S1Orbit.AS_BANDS:
                 raise ValueError(f"Cannot get S1 collection with {orbit=}")
-
+        col = col.map(f_mask_edges)
+        if self.speckle_filter_config:
+            col = speckle_filter_wrapper(col, *astuple(self.speckle_filter_config))
+        if self.terrain_normalization_config:
+            col = terrain_normalization_wrapper(col, *astuple(self.terrain_normalization_config))
         return col  # type: ignore[no-any-return]
 
     def get_time_series(
@@ -120,6 +135,8 @@ class S1(SatelliteABC):
         selected_bands: list[str] | None = None,
         resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
         resolution: float = 10,
+        speckle_filter_config: SpeckleFilterConfig | None = None,
+        terrain_normalization_config: TerrainNormalizationConfig | None = None,
         **kwargs: Any,
     ) -> DownloadableGeedimImageCollection:
         """Get a downloadable time series of Sentinel-1 images.
@@ -142,6 +159,8 @@ class S1(SatelliteABC):
             The resampling method to use when compositing images.
         resolution: float
             The resolution for the image.
+        speckle_filter_config : SpeckleFilterConfig | None
+        terrain_normalization_config : TerrainNormalizationConfig | None
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -150,10 +169,15 @@ class S1(SatelliteABC):
         s1_im: DownloadableGeedimImageCollection
             A Sentinel-1 time series collection of the specified AOI and time range.
         """
+        for key in kwargs:
+            if key not in ("speckle_filter_config", "terrain_normalization_config"):
+                log.warning(f"Argument {key} is ignored.")
+
+        self.speckle_filter_config = speckle_filter_config
+        self.terrain_normalization_config = terrain_normalization_config
         if orbit == S1Orbit.AS_BANDS:
             raise ValueError("Orbit AS_BANDS is not permitted for downloading time series.")
         s1_col = self.get_col(aoi, start_date, end_date, orbit, selected_bands)
-
         images = {}
         info = s1_col.getInfo()
         n_images = len(info["features"])  # type: ignore[index]
@@ -161,17 +185,24 @@ class S1(SatelliteABC):
             log.error(f"Found 0 Sentinel-1 image." f"Check region {aoi.transform(WGS84)}.")
             raise RuntimeError("Collection of 0 Sentinel-1 image.")
         for feature in info["features"]:  # type: ignore[index]
-            id_ = feature["id"]
-            footprint = PatchedBaseImage.from_id(id_).footprint
+            sys_index = feature.get("properties").get("system:index")
+            if self.is_preprocessed:
+                im = s1_col.filter(Filter.eq("system:index", sys_index)).first()
+                footprint = PatchedBaseImage.from_id(
+                    f"COPERNICUS/S1_GRD_FLOAT/{sys_index}"
+                ).footprint
+            else:
+                id_ = feature.get("id")
+                footprint = PatchedBaseImage.from_id(id_).footprint
+                im = Image(id_)
             assert footprint is not None
             if Polygon(footprint["coordinates"][0]).intersects(aoi.to_shapely_polygon()):
                 # aoi intersects im
-                im = Image(id_)
                 # convert to power and resample
                 im = self.before_composite(im, resampling, aoi, resolution)
                 # Apply pixel range and dtype
                 im = self.after_composite(im, dtype)
-                images[id_.removeprefix("COPERNICUS/S1_GRD/")] = PatchedBaseImage(im)
+                images[sys_index] = PatchedBaseImage(im)
         return DownloadableGeedimImageCollection(images)
 
     def get(
@@ -185,6 +216,8 @@ class S1(SatelliteABC):
         selected_bands: list[str] | None = None,
         resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
         resolution: float = 10,
+        speckle_filter_config: SpeckleFilterConfig | None = None,
+        terrain_normalization_config: TerrainNormalizationConfig | None = None,
         **kwargs: Any,
     ) -> DownloadableGeedimImage:
         """Get a downloadable mosaic of Sentinel-1 images.
@@ -209,6 +242,8 @@ class S1(SatelliteABC):
             The resampling method to use when compositing images.
         resolution: float
             The resolution for the image.
+        speckle_filter_config : SpeckleFilterConfig | None
+        terrain_normalization_config : TerrainNormalizationConfig | None
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -218,7 +253,11 @@ class S1(SatelliteABC):
             A Sentinel-1 composite image of the specified AOI and time range.
         """
         for key in kwargs:
-            log.warning(f"Argument {key} is ignored.")
+            if key not in ("speckle_filter_config", "terrain_normalization_config"):
+                log.warning(f"Argument {key} is ignored.")
+
+        self.speckle_filter_config = speckle_filter_config
+        self.terrain_normalization_config = terrain_normalization_config
 
         def get_im(orbit: S1Orbit) -> Image:
             s1_col = self.get_col(aoi, start_date, end_date, orbit, selected_bands)
@@ -267,8 +306,6 @@ class S1(SatelliteABC):
         aoi: GeoBoundingBox,
         scale: float,
     ) -> Image:
-        # Convert from db to power: 10^(im/10)
-        im = ee.Image(10).pow(im.divide(10))
         # Apply resampling if specified
         im = self.resample_reproject_clip(im, aoi, resampling, scale)
         return im