+++ tests/test_download.py
@@ -18,7 +18,7 @@ from geefetch.cli.download_implementation import (
     download_s1,
     download_s2,
 )
-from geefetch.cli.omegaconfig import load
+from geefetch.cli.omegaconfig import SpeckleFilterConfig, load
 from geefetch.data.process import tif_is_clean
 from geefetch.utils.enums import CompositeMethod, P2Orbit, ResamplingMethod, S1Orbit
 
@@ -35,6 +35,31 @@ def paris_config_path(
     conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
     yield conf_path
     conf_path.unlink()
+
+
+@pytest.fixture
+def paris_speckle_path(raw_paris_config: DictConfig, tmp_path: Path, gee_project_id: str):
+    raw_paris_config = raw_paris_config.copy()
+    raw_paris_config.data_dir = str(tmp_path)
+    raw_paris_config.satellite_default.gee.ee_project_id = gee_project_id
+    raw_paris_config.s1.speckle_filter = SpeckleFilterConfig()
+    conf_path = tmp_path / "config.yaml"
+    conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
+    return conf_path
+
+
+@pytest.fixture
+def paris_speckle_timeseries_path(
+    raw_paris_config: DictConfig, tmp_path: Path, gee_project_id: str
+):
+    raw_paris_config = raw_paris_config.copy()
+    raw_paris_config.data_dir = str(tmp_path)
+    raw_paris_config.satellite_default.gee.ee_project_id = gee_project_id
+    raw_paris_config.satellite_default.composite_method = CompositeMethod.TIMESERIES
+    raw_paris_config.s1.speckle_filter = SpeckleFilterConfig()
+    conf_path = tmp_path / "config.yaml"
+    conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
+    return conf_path
 
 
 @pytest.fixture
@@ -132,6 +157,26 @@ class TestDownloadSentinel1:
         download_s1(paris_timeseriesconfig_path)
         conf = load(paris_timeseriesconfig_path)
         downloaded_files = sorted(list(Path(conf.data_dir).rglob("*.tif")))
+        assert len(downloaded_files) == 5
+        assert downloaded_files[0].parts[-3:] == (
+            "s1",
+            "s1_EPSG2154_650000_6860000",
+            "S1A_IW_GRDH_1SDV_20200111T174030_20200111T174055_030756_0386DC_869A.tif",
+        )
+
+    def test_download_s1_with_speckle_filter(self, paris_speckle_path: Path):
+        download_s1(paris_speckle_path)
+        conf = load(paris_speckle_path)
+        downloaded_files = list(Path(conf.data_dir).rglob("*.tif"))
+        assert len(downloaded_files) == 1
+        assert downloaded_files[0].parts[-2:] == ("s1", "s1_EPSG2154_650000_6860000.tif")
+
+    def test_download_s1_with_speckle_filter_and_timeseries(
+        self, paris_speckle_timeseries_path: Path
+    ):
+        download_s1(paris_speckle_timeseries_path)
+        conf = load(paris_speckle_timeseries_path)
+        downloaded_files = list(Path(conf.data_dir).rglob("*.tif"))
         assert len(downloaded_files) == 5
         assert downloaded_files[0].parts[-3:] == (
             "s1",